back-end/requirements.txt
click==8.0.3
cryptography==36.0.1
Flask==2.0.2
Flask-Cors==3.0.10
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pycparser==2.21
pycryptodome==3.12.0
Werkzeug==2.0.2
